requirements.txt
SpeechRecognition
googletrans
pydub
webvtt-py
moviepy
ffmpeg-python
google-cloud-speech